ShareDialog: show shortened QRCode + refactor
